fix: update user entity and database helpers
Update user's entity `index.ts` exports
Fix FSD violation by moving method for getting user's DB data from helpers
Rename `updateUserInfo` to `updateUserData` to remove conflict with such selector
Rename `database.ts` to `deserializeFirebaseArray.ts`
