+++ src/entities/author/db/getAuthorData.ts
@@ -1,4 +1,4 @@
-import { deserializeFirebaseArray } from "@/shared/helpers/database";
+import { deserializeFirebaseArray } from "@/shared/helpers/deserializeFirebaseArray";
 import { getData } from "@/shared/lib/firebase";
 
 import { AuthorData } from "..";
