+++ src/entities/user/db/getUserData.ts
@@ -1,0 +1,22 @@
+import { deserializeFirebaseArray } from "@/shared/helpers/deserializeFirebaseArray";
+import { getData } from "@/shared/lib/firebase";
+
+import { UserDBData } from "../model/types";
+
+export const getUserData = async (userID: string) => {
+  try {
+    const dbPath = "users/" + userID;
+    const userData = await getData<UserDBData>(dbPath);
+
+    const deserializedData = {
+      ...userData,
+      tweetsIds: deserializeFirebaseArray(userData.tweetsIds),
+      followersIds: deserializeFirebaseArray(userData.followersIds),
+      followingIds: deserializeFirebaseArray(userData.followingIds),
+    };
+
+    return deserializedData;
+  } catch (error) {
+    throw new Error(`Failed to fetch user data! ${(error as Error).message}`);
+  }
+};
