+++ src/features/profile/db/uploadUserAvatar.ts
@@ -1,6 +1,6 @@
 import { getDownloadURL, ref, uploadBytes } from "firebase/storage";
 
-import { updateUserInfo } from "@/entities/user/db/updateUserInfo";
+import { updateUserData } from "@/entities/user";
 import { storage } from "@/shared/lib/firebase";
 
 type FileDataFormats = Blob | Uint8Array | ArrayBuffer;
@@ -13,7 +13,7 @@ export const uploadUserAvatar = async (
     const userAvatarRef = ref(storage, `users/${userId}/avatar-image`);
     const uploadResult = await uploadBytes(userAvatarRef, newAvatar);
     const avatarURL = await getDownloadURL(uploadResult.ref);
-    await updateUserInfo(userId, {
+    await updateUserData(userId, {
       avatarURL,
     });
     return avatarURL;
