+++ src/shared/helpers/deserializeFirebaseArray.ts
@@ -1,0 +1,4 @@
+export const deserializeFirebaseArray = <T>(data: FirebaseArrayValue<T>) => {
+  if (data === undefined) return [];
+  return Object.values(data);
+};
